Create logic/schema.py and validators.py
